Run e2e test suite in CI.
