+++ .github/workflows/integration-tests.yml
@@ -11,11 +11,8 @@ jobs:
     runs-on: ubuntu-latest
     # From https://docs.github.com/en/actions/tutorials/use-containerized-services/create-postgresql-service-containers#configuring-the-runner-job-for-jobs-directly-on-the-runner-machine
     services:
-      # Label used to access the service container
       postgres:
-        # Docker Hub image
         image: postgres
-        # Provide the password for postgres
         env:
           POSTGRES_PASSWORD: postgres
         # Set health checks to wait until postgres has started
@@ -38,7 +35,10 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
 
-      - name: Install dependencies
+      - name: Install system dependencies
+        run: sudo apt update && sudo apt install -y xvfb ffmpeg tmux firefox coreutils
+
+      - name: Install python dependencies
         run: python -m pip install .[drf]
 
       - name: Install dependencies
@@ -46,10 +46,18 @@ jobs:
           -r requirements/requirements-dev.in
           -r requirements/requirements-test.in
 
-      - name: Run unit tests
-        run: python3 run_e2e_tests.py
+      - name: Run e2e tests
+        run: ./ci_run_e2e_tests.sh
         env:
           # The hostname used to communicate with the PostgreSQL service container
           POSTGRES_HOST: localhost
           # The default PostgreSQL port
           POSTGRES_PORT: 5432
+
+      - name: Archive tests artifacts
+        uses: actions/upload-artifact@v4
+        with:
+          name: log_and_video
+          path: |
+            *.log
+            *.mp4
